Extracted nested editor util functions
no issue

- pulled the new `setupNestedEditor` and `populateNestedEditor` functions out of `ProductNode` ready for re-use in other nodes
- modified `setupNestedEditor()` and `generateEditorState()` to use the `history-merge` tag to avoid unwanted undo states related to the initial editor setup/population
